Deduplicate ST0 GetPlayerCollisionWith
