+++ config/symbols.us.stdre.txt
@@ -15,7 +15,6 @@ g_eBreakableanimSets = 0x80180640;
 g_eBreakableDrawModes = 0x80180650;
 g_CloneShootOrder = 0x801807D4;
 UNK_Invincibility0 = 0x8018097C;
-g_SineTable = 0x80180D80;
 g_eUnk14SpawnRots = 0x801812D4;
 unk15_rot = 0x801812E4;
 unk15_yVel = 0x801812F4;
