+++ src/st/nz0/st_common.c
@@ -1,4 +1,3 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "nz0.h"
 #include "../st_common.h"
-
