+++ src/st/st0/st_common.c
@@ -113,92 +113,6 @@ static s16 g_SineTable[] = {
 
 #include "../check_field_collision.h"
 
-// This function checks if the player collides with the specified self
-// and from which direction.
-// w and h holds the collider size of the self
-// while flags stores which sides are solid
-s32 GetPlayerCollisionWith(Entity* self, u16 w, u16 h, u16 flags) {
-    Entity* pl;
-    s16 x;
-    s16 y;
-    u16 checks;
-
-    pl = &PLAYER;
-
-    if (self->posX.i.hi & 0x100) {
-        return 0;
-    }
-    if (self->posY.i.hi & 0x100) {
-        return 0;
-    }
-
-    x = pl->posX.i.hi - self->posX.i.hi;
-    y = pl->posY.i.hi - self->posY.i.hi;
-    if (self->facingLeft) {
-        x += self->hitboxOffX;
-    } else {
-        x -= self->hitboxOffX;
-    }
-    y -= self->hitboxOffY;
-
-    if (x > 0) {
-        checks = 1;
-    } else {
-        checks = 0;
-    }
-    if (y > 0) {
-        checks |= 2;
-    }
-
-    w += 8;
-    h += 24;
-
-    x += w;
-    y += h;
-    w += w;
-    h += h;
-    if ((u16)x > w || (u16)y > h) {
-        return 0;
-    }
-
-    if (x && x != w) {
-        // check collision from top
-        if (flags & 4 && checks ^ 2 && pl->velocityY >= 0 && y < 8) {
-            pl->posY.i.hi -= y;
-            g_Player.pl_vram_flag |= 0x41;
-            return 4;
-        }
-
-        // check collision from bottom
-        if (flags & 2 && checks & 2 && (pl->velocityY <= 0 || flags & 0x10)) {
-            y = h - y;
-            if (y < 0x10) {
-                pl->posY.i.hi += y;
-                g_Player.pl_vram_flag |= 0x42;
-                return 2;
-            }
-        }
-    }
-
-    // check collision from the sides
-    if (y && y != h && flags & 1) {
-        if (checks & 1) {
-            x = w - x;
-            if (flags & 8 && x > 2) {
-                x = 2;
-            }
-            pl->posX.i.hi += x;
-            return 1;
-        } else {
-            if (flags & 8 && x > 2) {
-                x = 2;
-            }
-            pl->posX.i.hi -= x;
-            return 1;
-        }
-    }
-
-    return 0;
-}
+#include "../get_player_collision_with.h"
 
 #include "../replace_breakable_with_item_drop.h"
